k8s: Add SYS_MODULE capability, module paths
Add kernel module autoload capability, needed for IPv6 TPROXY in cases
where the tables are not yet loaded when cilium-agent starts.

Also add volume mounts for /lib/modules and /sbin/modprobe, as
'ip[6]tables' command depends on these when a kernel modules is not
already loaded.

Note that these are also needed for IPSec NOTRACK rules.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>