identity: Change globalIdentity to wrap a LabelArray
Change the identity/cache package internal type 'globalIdentity' to
wrap a (sorted) LabelArray rather than (unsorted) Labels, as that is a
better fit for (always sorted) KV-store keys.

Add the ID and Key fields to the events generated by the local
identity allocations so that the watcher can know what has
happened. The same globalIdentity type is used for this purpose as
well (this type could be renamed to cacheIdentity?).

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>