cilium: ipsec_linux only set spi bit in xfrm mark on egress
When we add egress encryption policies via xfrm we add them using a
mark/mask value like this (mark=0xke00 mask=0xff00) where 'k' is the
Key index referencing the associated keys. This is used to select the
key to use for encryption on egress when multiple keys are in-use
during key rotation or encryption enablement/disablement. For example
(mark=0x1e00) will only apply to skbs that should be encrypted with
keys added with the 1 index.  Similarly 0x2e00 will encrypt with the
keys that have index 2, and so on.

Being a bit lazy perhaps we install decryption policies using the same
mark but with a mask the 'k' value masked out, e.g. (mark=0xkd00 mask=0x0f00)
because the key to use for decryption is specified by the SPI field in
the ESP header. Parsing the packet in BPF to additionally provide the
'k' is wasted work.

However, when we attempt to delete the keys the kernel compares policy
entries to find a matching policy to delete like so,

		if (pol->type == type &&
		    (mark & pol->mark.m) == pol->mark.v &&
		    !selector_cmp(sel, &pol->selector) &&
		    xfrm_sec_ctx_match(ctx, pol->security)) {

Where 'mark' is the value passed in by user, 'pol->mark.m' is the
installed policies mask value and 'pol->mark.v' is the installed
policies mark value. So what is happening is the decrypt entries (the
ones with mark=0xkd00 mask=0x0f00) get stuck in the policy database and
can never be deleted. This happens because the policies mask is applied
to the input before comparing with the value. This also explains why the
key rollover logic gets stuck sometimes because old policies are not
being removed completely.

This seems like a kernel bug or at least a strange design but the fix
from user space is to simple. After this patch the decrypt rules do not
include the 'k' bits so that (mark=0x0d00 mask0x0f00). After which the
delete operation is sucessful.

This bug was introduced when we added (re)keying support.

Unit tests worked previously because if we don't try to add any
conflicting entries nothing "bad" happens but this PR deletes
policy/state to add a new policy/state with the same src/dst
mark/mask but with new keys. And the failure above is seen.

Fixes: b6989723a7cc ("cilium: ipsec, support rolling updates")
Signed-off-by: John Fastabend <john.fastabend@gmail.com>