cilium: ipsec tests should use decodeIPSecKey for strings to hex
Currently, we pass keys in to byte[] as a string which is not reliable
it seems to work but that is just by chance. Use the decodeIPSecKey()
routine to convert strings to hex keys.

Note, all code but tests use the decodeIPSecKey() so this also brings
our test code inline with daemon code.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>